Remove notion of custom actions
Now, we only allow customization through action steps
