+++ jbi/runner.py
@@ -6,12 +6,15 @@ from typing import Optional
 
 from statsd.defaults.env import statsd
 
-from jbi import Operation
-from jbi.environment import Settings
+from jbi import ActionResult, Operation
+from jbi import steps as steps_module
+from jbi.environment import get_settings
 from jbi.errors import ActionNotFoundError, IgnoreInvalidRequestError
 from jbi.models import (
     ActionContext,
+    ActionParams,
     Actions,
+    ActionSteps,
     BugzillaWebhookComment,
     BugzillaWebhookRequest,
     JiraContext,
@@ -21,12 +24,71 @@ from jbi.services import bugzilla, jira
 
 logger = logging.getLogger(__name__)
 
+settings = get_settings()
+
+
+GROUP_TO_OPERATION = {
+    "new": Operation.CREATE,
+    "existing": Operation.UPDATE,
+    "comment": Operation.COMMENT,
+}
+
+
+def groups2operation(steps: ActionSteps):
+    """In the configuration files, the steps are grouped by `new`, `existing`,
+    and `comment`. Internally, this correspond to enums of `Operation`.
+    This helper remaps the list of steps.
+    """
+    try:
+        by_operation = {
+            GROUP_TO_OPERATION[entry]: steps_list
+            for entry, steps_list in steps.dict().items()
+        }
+    except KeyError as err:
+        raise ValueError(f"Unsupported entry in `steps`: {err}") from err
+    return by_operation
+
+
+class Executor:
+    """Callable class that encapsulates the default action."""
+
+    def __init__(self, parameters: ActionParams):
+        self.parameters = parameters
+        self.steps = self._initialize_steps(parameters.steps)
+
+    def _initialize_steps(self, steps: ActionSteps):
+        steps_by_operation = groups2operation(steps)
+        steps_callables = {
+            group: [getattr(steps_module, step_str) for step_str in steps_list]
+            for group, steps_list in steps_by_operation.items()
+        }
+        return steps_callables
+
+    def __call__(self, context: ActionContext) -> ActionResult:
+        """Called from `runner` when the action is used."""
+
+        responses = tuple()  # type: ignore
+
+        for step in self.steps[context.operation]:
+            context, step_responses = step(context=context, parameters=self.parameters)
+            for response in step_responses:
+                logger.debug(
+                    "Received %s",
+                    response,
+                    extra={
+                        "response": response,
+                        **context.dict(),
+                    },
+                )
+            responses += step_responses
+
+        return True, {"responses": responses}
+
 
 @statsd.timer("jbi.action.execution.timer")
 def execute_action(
     request: BugzillaWebhookRequest,
     actions: Actions,
-    settings: Settings,
 ):
     """Execute the configured action for the specified `request`.
 
@@ -111,14 +173,13 @@ def execute_action(
             )
 
         logger.info(
-            "Execute action '%s:%s' for Bug %s",
+            "Execute action '%s' for Bug %s",
             action.whiteboard_tag,
-            action.module,
             bug.id,
             extra=runner_context.update(operation=Operation.EXECUTE).dict(),
         )
-
-        handled, details = action.caller(context=action_context)
+        executor = Executor(parameters=action.parameters)
+        handled, details = executor(context=action_context)
 
         logger.info(
             "Action %r executed successfully for Bug %s",
